refactor: consolidate Graph API services into Infrastructure/Graph folder
